feat: remove old switcher and rename component
